Test refactoring: move QueryTest's pymongo connection into a base class
